Restructure repository structure for Stac modules
