+++ src/geospatial_tools/stac/copernicus/__init__.py
@@ -1,6 +1,6 @@
 """Copernicus Data Space Ecosystem (CDSE) tools and constants."""
 
-from .sentinel_2 import (
+from geospatial_tools.stac.copernicus.constants import (
     CopernicusS2Band,
     CopernicusS2Collection,
     CopernicusS2Resolution,
