+++ src/geospatial_tools/stac/copernicus/auth.py
