+++ src/geospatial_tools/stac/copernicus/constants.py
